norm now long

+++ src/main/java/de/gupta/commons/utility/math/algebra/element/ring/concrete/IntegerRing.java
@@ -47,7 +47,7 @@ public record IntegerRing(int value) implements EuclideanDomain<IntegerRing>
 	}
 
 	@Override
-	public int norm()
+	public long norm()
 	{
 		return Math.absExact(value);
 	}
