+++ src/test/java/de/gupta/commons/utility/math/algebra/element/ring/RingElementDefaultsTest.java
@@ -185,7 +185,7 @@ class RingElementDefaultsTest
 		}
 
 		@Override
-		public int norm()
+		public long norm()
 		{
 			return Math.abs(value);
 		}
